Extract SortTags to a util
